# Conflicts:
#	public/js/reservation/PMS0110040.js
#	public/src/reservation/PMS0110040/components/PMS0110041_LITE.vue

+++ public/src/setup/PMS0810230/components/useTime.vue
@@ -12,7 +12,7 @@
                                         row-click-color="#edf7ff"
                                         is-horizontal-resize
                                         style="width:100%"
-                                        :columns="useTimeColumns"
+                                        :columns="useTimeColumns" _
                                         :table-data="useTimeData"
                                         :error-content="errorContent"
                                         :column-cell-class-name="useTimeColumnCellClass"
@@ -33,17 +33,22 @@
                                     <li>
                                         <button class="btn btn-primary btn-white btn-defaultWidth"
                                                 role="button" @click="confirmData">{{i18nLang.program.PMS0810230.OK}}
+
+
                                         </button>
                                     </li>
                                     <li>
                                         <button class="btn btn-primary btn-white btn-defaultWidth"
                                                 role="button" @click="closeDialog">{{i18nLang.program.PMS0810230.leave}}
+
+
                                         </button>
                                     </li>
                                     <li>
                                         <span class="checkbox">
                                               <label class="checkbox-width">
-                                                  <input name="form-field-checkbox" type="checkbox" class="ace" v-model="isShowExpire" @change="showTable">
+                                                  <input name="form-field-checkbox" type="checkbox" class="ace"
+                                                         v-model="isShowExpire" @change="showTable">
                                                   <span class="lbl font-btn">{{i18nLang.program.PMS0810230.showExpire}}</span>
                                               </label>
                                           </span>
@@ -86,15 +91,22 @@
         },
         methods: {
             deleteRow() {
-                let params = {type: 'delete', index: this.index, rowData: this.rowData};
-                this.$emit('on-custom-comp', params);
+                let lb_isReadOnly = false;
+                //開始日已過滾房租日期時開始日欄位唯讀不可修改
+                lb_isReadOnly = compareRentCalDat(this.rowData.begin_dat, this.rowData.rentCalDat);
+                //當筆使用期間結束日期大於滾房租日期時為整筆唯讀，不可修改。
+                lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                if (!lb_isReadOnly) {
+                    let params = {type: 'delete', index: this.index, rowData: this.rowData};
+                    this.$emit('on-custom-comp', params);
+                }
             }
         }
     });
     //起始日
     Vue.component('table-begin-date', {
-        template: '<el-date-picker v-model="rowData.begin_dat" type="date" format="yyyy/MM/dd"' +
-        'style="width: 135px; height: 35px;line-height: 25px;" editable="false" @change="changeRowData(rowData, index)" :readonly="rowData.isEdit"' +
+        template: '<el-date-picker v-model="rowData.begin_dat" type="date" format="yyyy/MM/dd" :class="{input_sta_required: !isReadOnly}"' +
+        'style="width: 135px; height: 35px;line-height: 25px;" editable="false" @change="changeRowData(rowData, index)" :readonly="isReadOnly"' +
         ':editable="false" :clearable="false"></el-date-picker>',
         props: {
             rowData: {
@@ -104,9 +116,23 @@
                 type: Number
             }
         },
+        computed: {
+            isReadOnly() {
+                let lb_isReadOnly = false;
+                lb_isReadOnly = compareRentCalDat(this.rowData.begin_dat, this.rowData.rentCalDat);
+                //當筆使用期間結束日期大於滾房租日期時為整筆唯讀，不可修改。
+                lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                return lb_isReadOnly
+            }
+        },
         watch: {
             "rowData.begin_dat": function (val, oldVal) {
-                if (moment(val).diff(moment(this.rowData.end_dat), "days") > 0) {
+                //chk_1010_begin_dat
+                if (moment(val).diff(moment(this.rowData.rentCalDat), "days") <= 0) {
+                    alert(go_i18nLang.program.PMS0810230.begBiggerEnd);
+                    this.rowData.begin_dat = oldVal;
+                }
+                else if (moment(val).diff(moment(this.rowData.end_dat), "days") > 0) {
                     alert(go_i18nLang.program.PMS0810230.begBiggerEnd);
                     this.rowData.begin_dat = oldVal;
                 }
@@ -122,8 +148,8 @@
     });
     //結束日
     Vue.component('table-end-date', {
-        template: '<el-date-picker v-model="rowData.end_dat" type="date" format="yyyy/MM/dd" class="input_sta_required "' +
-        'style="width: 135px; height: 35px; line-height: 25px;"  @change="changeRowData(rowData, index)"' +
+        template: '<el-date-picker v-model="rowData.end_dat" type="date" format="yyyy/MM/dd" :class="{input_sta_required: !isReadOnly}" ' +
+        'style="width: 135px; height: 35px; line-height: 25px;"  @change="changeRowData(rowData, index)"  :readonly="isReadOnly"' +
         ':editable="false" :clearable="false"></el-date-picker>',
         props: {
             rowData: {
@@ -131,6 +157,12 @@
             },
             index: {
                 type: Number
+            }
+        },
+        computed: {
+            isReadOnly() {
+                let lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                return lb_isReadOnly
             }
         },
         watch: {
@@ -152,7 +184,7 @@
     //計算方式
     Vue.component('table-command-cod', {
         template: '<bac-select v-model="rowData.command_cod" :field="rowData.cmFieldsData" :data="rowData.cmFieldsData.selectData" ' +
-        ':data-display="rowData.cmFieldsData.selectData" style="width: 135px; height: 25px;"' +
+        ':data-display="rowData.cmFieldsData.selectData" style="width: 135px; height: 25px;" :readonly="isReadOnly"' +
         ':default-val="rowData.command_cod" is-qry-src-before="Y" value-field="value" text-field="display"' +
         '@update:v-model="val => rowData.command_cod = val" @change="changeRowData(rowData, index)"></bac-select>',
         props: {
@@ -164,6 +196,12 @@
             },
             index: {
                 type: Number
+            }
+        },
+        computed: {
+            isReadOnly() {
+                let lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                return lb_isReadOnly
             }
         },
         watch: {
@@ -184,7 +222,7 @@
     //日其規則
     Vue.component('table-command-option', {
         template: '<bac-select v-model="rowData.command_option" :field="rowData.dtdFieldsData" :data="rowData.dtdFieldsData.selectData" ' +
-        ':data-display="rowData.dtdFieldsData.selectDataDisplay" style="width: 135px; height: 25px;" :multiple="true"' +
+        ':data-display="rowData.dtdFieldsData.selectDataDisplay" style="width: 135px; height: 25px;" :multiple="true" :readonly="isReadOnly"' +
         ':default-val="rowData.command_option" is-qry-src-before="Y" value-field="value" text-field="display"' +
         '@update:v-model="val => rowData.command_option = val" @change="changeRowData(rowData, index)"></bac-select>',
         props: {
@@ -196,6 +234,12 @@
             },
             index: {
                 type: Number
+            }
+        },
+        computed: {
+            isReadOnly() {
+                let lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                return lb_isReadOnly
             }
         },
         watch: {
@@ -225,12 +269,11 @@
                 this.$emit('on-custom-comp', lo_params);
             }
         }
-    })
-    ;
+    });
     //房型
     Vue.component('table-room-cods', {
         template: '<bac-select v-model="rowData.room_cods" :field="rowData.rcFieldsData" :data="rowData.rcFieldsData.selectData" ' +
-        ':data-display="rowData.rcFieldsData.selectData" style="width: 135px; height: 25px;" :multiple="true"' +
+        ':data-display="rowData.rcFieldsData.selectData" style="width: 135px; height: 25px;" :multiple="true" :readonly="isReadOnly"' +
         ':default-val="rowData.room_cods" is-qry-src-before="Y" value-field="value" text-field="display"' +
         '@update:v-model="val => rowData.room_cods = val" @change="changeRowData(rowData, index)"></bac-select>',
         props: {
@@ -242,6 +285,12 @@
             },
             index: {
                 type: Number
+            }
+        },
+        computed: {
+            isReadOnly() {
+                let lb_isReadOnly = compareRentCalDat(this.rowData.end_dat, this.rowData.rentCalDat);
+                return lb_isReadOnly
             }
         },
         watch: {
@@ -422,7 +471,6 @@
                         _.each(la_dgRowData, (lo_dgRowData, idx) => {
                             la_dgRowData[idx]["begin_dat"] = moment(lo_dgRowData["begin_dat"]).format("YYYY/MM/DD");
                             la_dgRowData[idx]["end_dat"] = moment(lo_dgRowData["end_dat"]).format("YYYY/MM/DD");
-                            la_dgRowData[idx]["isEdit"] = moment(lo_dgRowData["end_dat"]).format("YYYY/MM/DD");
                         });
 
                         this.fieldsData = result.dgFieldsData;
@@ -506,29 +554,25 @@
                     this.useTimeColumns.splice(3, 0, lo_commandCod);
                 }
                 this.useTimeData = [];
-                let la_displayDataGridRowsData = this.isShowExpire ?
-                    this.dataGridRowsData : _.filter(this.dataGridRowsData, (lo_dataGridRowsData) => {
-                        let lo_endDat = moment(lo_dataGridRowsData.end_dat);
-                        let lo_rentCalDat = moment(this.rentCalDat);
-                        return lo_endDat.diff(lo_rentCalDat, 'days') >= 1
+                if (this.dataGridRowsData.length > 0) {
+                    _.each(this.dataGridRowsData, (lo_dataGridRowsData) => {
+                        let lb_overDate = moment(lo_dataGridRowsData.end_dat).diff(this.rentCalDat, 'days') >= 1 ? false : true;
+                        if (this.isShowExpire || !this.isShowExpire && !lb_overDate) {
+                            this.useTimeData.push({
+                                "begin_dat": moment(lo_dataGridRowsData.begin_dat).format("YYYY/MM/DD"),
+                                "end_dat": moment(lo_dataGridRowsData.end_dat).format("YYYY/MM/DD"),
+                                "command_cod": lo_dataGridRowsData.command_cod,
+                                "command_option": _.isUndefined(lo_dataGridRowsData.command_option) ? "" : this.convertSelectData(lo_dataGridRowsData.command_option),
+                                "room_cods": _.isUndefined(lo_dataGridRowsData.room_cods) ? "" : this.convertSelectData(lo_dataGridRowsData.room_cods),
+                                "supply_nos": lo_dataGridRowsData.supply_nos,
+                                "cmFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'command_cod'}),
+                                "dtdFieldsData": this.convertCommandOption(lo_dataGridRowsData),
+                                "rcFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'room_cods'}),
+                                "rentCalDat": this.rentCalDat
+                            });
+                        }
                     });
-                if (la_displayDataGridRowsData.length > 0) {
-                    _.each(la_displayDataGridRowsData, (lo_dataGridRowsData) => {
-                        this.useTimeData.push({
-                            "begin_dat": moment(lo_dataGridRowsData.begin_dat).format("YYYY/MM/DD"),
-                            "end_dat": moment(lo_dataGridRowsData.end_dat).format("YYYY/MM/DD"),
-                            "command_cod": lo_dataGridRowsData.command_cod,
-                            "command_option": this.convertSelectData(lo_dataGridRowsData.command_option),
-                            "room_cods": _.isUndefined(lo_dataGridRowsData.room_cods) ? "" : this.convertSelectData(lo_dataGridRowsData.room_cods),
-                            "supply_nos": lo_dataGridRowsData.supply_nos,
-                            "cmFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'command_cod'}),
-                            "dtdFieldsData": this.convertCommandOption(lo_dataGridRowsData),
-                            "rcFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'room_cods'}),
-                            "isEdit": _.isUndefined(lo_dataGridRowsData.isEdit) ? false : true
-                        });
-                    });
-                }
-                else {
+                } else {
                     this.useTimeData = [{
                         "cmFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'command_cod'}),
                         "dtdFieldsData": _.findWhere(this.fieldsData, {ui_field_name: 'command_option'}),
@@ -668,7 +712,7 @@
                         rate_cod: this.$store.state.gs_rateCod,
                         supply_nos: Number(ln_maxSupplyNos) + 1,
                         begin_dat: moment().format("YYYY/MM/DD"),
-                        end_dat: moment().add(1, "day").format("YYYY/MM/DD"),
+                        end_dat: moment().endOf('year').format("YYYY/MM/DD"),
                         command_cod: "H",
                         command_option: la_commandOptionHSelect.length > 0 ? _.first(la_commandOptionHSelect).value : "",
                         room_cods: la_roomCosSelect.length > 0 ? _.first(la_roomCosSelect).value : "",
@@ -680,6 +724,7 @@
                 }
             },
             async doValidate() {
+                //TODO 將此規則回填至SD
                 let lo_checkResult = {success: true, msg: ""};
 
                 if (this.chgDataGridRowsData.length == 0) {
@@ -811,5 +856,9 @@
                 return la_returnData;
             }
         }
+    }
+
+    function compareRentCalDat(nowDate, rentCalDat) {
+        return moment(nowDate).diff(rentCalDat) < 0 ? true : false;
     }
 </script>
